add admin event  message handler

+++ chat/src/main/java/io/projectZ/orchestrator/infrastructure/adapter/in/broker/handler/UserEventHandler.java
@@ -1,4 +1,4 @@
-package io.projectZ.orchestrator.infrastructure.adapter.in.broker;
+package io.projectZ.orchestrator.infrastructure.adapter.in.broker.handler;
 /*
   Project : Orchestrator
   Author  : AmirHFF
@@ -7,12 +7,13 @@ package io.projectZ.orchestrator.infrastructure.adapter.in.broker;
 
 import io.projectZ.orchestrator.application.service.ProfileService;
 import io.projectZ.orchestrator.entity.ChatProfile;
+import io.projectZ.orchestrator.infrastructure.adapter.in.broker.dto.AdminProfileEventDto;
 import io.projectZ.orchestrator.infrastructure.adapter.in.broker.dto.EventDTO;
-import io.projectZ.orchestrator.infrastructure.adapter.in.broker.dto.UserEventDto;
+import io.projectZ.orchestrator.infrastructure.adapter.in.broker.dto.UserProfileEventDto;
 import org.springframework.stereotype.Component;
 
-@Component
-public class UserEventHandler implements EventHandler<UserEventDto> {
+@Component("userEvent")
+public class UserEventHandler implements EventHandler {
 
     private ProfileService profileService;
 
@@ -21,16 +22,18 @@ public class UserEventHandler implements EventHandler<UserEventDto> {
     }
 
     @Override
-    public void handle(UserEventDto userEventDto) {
-        ChatProfile chatProfile = mapUserEventToChatProfile(userEventDto);
-        switch (userEventDto.getEventType()){
-            case REGISTER -> profileService.saveProfile(chatProfile);
-            case UPDATE_PROFILE -> profileService.updateProfile(chatProfile);
-            case DELETE_ACCOUNT -> profileService.remove(chatProfile.getUsername());
+    public void handle(EventDTO eventDTO) {
+        if (eventDTO instanceof UserProfileEventDto userEventDto) {
+            ChatProfile chatProfile = mapUserEventToChatProfile(userEventDto);
+            switch (userEventDto.getEventType()) {
+                case REGISTER -> profileService.saveProfile(chatProfile);
+                case UPDATE_PROFILE -> profileService.updateProfile(chatProfile);
+                case DELETE_ACCOUNT -> profileService.remove(chatProfile.getUsername());
+            }
         }
     }
 
-    private ChatProfile mapUserEventToChatProfile(UserEventDto userEventDto){
+    private ChatProfile mapUserEventToChatProfile(UserProfileEventDto userEventDto) {
         ChatProfile chatProfile = new ChatProfile();
         chatProfile.setUsername(userEventDto.getDetails().get("username"));
         chatProfile.setFirstName(userEventDto.getDetails().get("first_name"));
